+++ sdk/keyvault/azure-keyvault-secrets/azure/keyvault/secrets/aio/_client.py
@@ -7,12 +7,11 @@ from functools import partial
 
 from azure.core.tracing.decorator import distributed_trace
 from azure.core.tracing.decorator_async import distributed_trace_async
-from azure.core.polling import async_poller
 
 from .._models import KeyVaultSecret, DeletedSecret, SecretProperties
 from .._shared import AsyncKeyVaultClientBase
 from .._shared.exceptions import error_map as _error_map
-from .._shared._polling_async import DeleteAsyncPollingMethod, RecoverDeletedAsyncPollingMethod
+from .._shared._polling_async import AsyncDeleteRecoverPollingMethod
 
 
 class SecretClient(AsyncKeyVaultClientBase):
@@ -88,9 +87,7 @@ class SecretClient(AsyncKeyVaultClientBase):
         not_before = kwargs.pop("not_before", None)
         expires_on = kwargs.pop("expires_on", None)
         if enabled is not None or not_before is not None or expires_on is not None:
-            attributes = self._models.SecretAttributes(
-                enabled=enabled, not_before=not_before, expires=expires_on
-            )
+            attributes = self._models.SecretAttributes(enabled=enabled, not_before=not_before, expires=expires_on)
         else:
             attributes = None
         bundle = await self._client.set_secret(
@@ -132,9 +129,7 @@ class SecretClient(AsyncKeyVaultClientBase):
         not_before = kwargs.pop("not_before", None)
         expires_on = kwargs.pop("expires_on", None)
         if enabled is not None or not_before is not None or expires_on is not None:
-            attributes = self._models.SecretAttributes(
-                enabled=enabled, not_before=not_before, expires=expires_on
-            )
+            attributes = self._models.SecretAttributes(enabled=enabled, not_before=not_before, expires=expires_on)
         else:
             attributes = None
         bundle = await self._client.update_secret(
@@ -265,13 +260,17 @@ class SecretClient(AsyncKeyVaultClientBase):
         deleted_secret = DeletedSecret._from_deleted_secret_bundle(
             await self._client.delete_secret(self.vault_url, name, error_map=_error_map, **kwargs)
         )
-        sd_disabled = deleted_secret.recovery_id is None
-        command = partial(self.get_deleted_secret, name=name, **kwargs)
 
-        delete_secret_poller = DeleteAsyncPollingMethod(
-            initial_status="deleting", finished_status="deleted", sd_disabled=sd_disabled, interval=polling_interval
+        polling_method = AsyncDeleteRecoverPollingMethod(
+            # no recovery ID means soft-delete is disabled, in which case we initialize the poller as finished
+            command=partial(self.get_deleted_secret, name=name, **kwargs),
+            final_resource=deleted_secret,
+            finished=deleted_secret.recovery_id is None,
+            interval=polling_interval,
         )
-        return await async_poller(command, deleted_secret, None, delete_secret_poller)
+        await polling_method.run()
+
+        return polling_method.resource()
 
     @distributed_trace_async
     async def get_deleted_secret(self, name: str, **kwargs: "Any") -> DeletedSecret:
@@ -371,9 +370,11 @@ class SecretClient(AsyncKeyVaultClientBase):
         recovered_secret = SecretProperties._from_secret_bundle(
             await self._client.recover_deleted_secret(self.vault_url, name, error_map=_error_map, **kwargs)
         )
-        command = partial(self.get_secret, name=name, **kwargs)
 
-        recover_secret_poller = RecoverDeletedAsyncPollingMethod(
-            initial_status="recovering", finished_status="recovered", interval=polling_interval
+        command = partial(self.get_secret, name=name, **kwargs)
+        polling_method = AsyncDeleteRecoverPollingMethod(
+            command=command, final_resource=recovered_secret, finished=False, interval=polling_interval
         )
-        return await async_poller(command, recovered_secret, None, recover_secret_poller)
+        await polling_method.run()
+
+        return polling_method.resource()