+++ sdk/servicebus/azure-servicebus/tests/servicebus_preparer.py
@@ -40,6 +40,8 @@ class ServiceBusNamespacePreparer(AzureMgmtPreparer):
         self.resource_group_parameter_name = resource_group_parameter_name
         self.parameter_name = parameter_name
         self.connection_string = ''
+        if random_name_enabled:
+            self.resource_moniker = self.name_prefix + "sbname"
 
     def create_resource(self, name, **kwargs):
         if self.is_live:
@@ -59,6 +61,11 @@ class ServiceBusNamespacePreparer(AzureMgmtPreparer):
             self.connection_string = key.primary_connection_string
             self.key_name = key.key_name
             self.primary_key = key.primary_key
+
+            self.test_class_instance.scrubber.register_name_pair(
+                name,
+                self.resource_moniker
+            )
         else:
             self.resource = FakeResource(name=name, id=name)
             self.connection_string = 'Endpoint=sb://test-azure-sdk-test.servicebus.windows.net/;SharedAccessKeyName=test;SharedAccessKey=THISISATESTKEYXXXXXXXXXXXXXXXXXXXXXXXXXXXX='
@@ -134,6 +141,8 @@ class ServiceBusTopicPreparer(_ServiceBusChildResourcePreparer):
                                                      playback_fake_resource=playback_fake_resource,
                                                      client_kwargs=client_kwargs)
         self.parameter_name = parameter_name
+        if random_name_enabled:
+            self.resource_moniker = self.name_prefix + "sbtopic"
 
     def create_resource(self, name, **kwargs):
         if self.is_live:
@@ -145,6 +154,11 @@ class ServiceBusTopicPreparer(_ServiceBusChildResourcePreparer):
                 namespace.name,
                 name,
                 {}
+            )
+
+            self.test_class_instance.scrubber.register_name_pair(
+                name,
+                self.resource_moniker
             )
         else:
             self.resource = FakeResource(name=name, id=name)
@@ -178,6 +192,8 @@ class ServiceBusSubscriptionPreparer(_ServiceBusChildResourcePreparer):
                                                      client_kwargs=client_kwargs)
         self.servicebus_topic_parameter_name = servicebus_topic_parameter_name
         self.parameter_name = parameter_name
+        if random_name_enabled:
+            self.resource_moniker = self.name_prefix + "sbsub"
 
     def create_resource(self, name, **kwargs):
         if self.is_live:
@@ -191,6 +207,11 @@ class ServiceBusSubscriptionPreparer(_ServiceBusChildResourcePreparer):
                 topic.name,
                 name,
                 {}
+            )
+
+            self.test_class_instance.scrubber.register_name_pair(
+                name,
+                self.resource_moniker
             )
         else:
             self.resource = FakeResource(name=name, id=name)
@@ -240,6 +261,8 @@ class ServiceBusQueuePreparer(_ServiceBusChildResourcePreparer):
         self.requires_duplicate_detection=requires_duplicate_detection
         self.dead_lettering_on_message_expiration=dead_lettering_on_message_expiration
         self.requires_session=requires_session
+        if random_name_enabled:
+            self.resource_moniker = self.name_prefix + "sbqueue"
 
     def create_resource(self, name, **kwargs):
         if self.is_live:
@@ -255,6 +278,11 @@ class ServiceBusQueuePreparer(_ServiceBusChildResourcePreparer):
                     requires_duplicate_detection = self.requires_duplicate_detection,
                     dead_lettering_on_message_expiration = self.dead_lettering_on_message_expiration,
                     requires_session = self.requires_session)
+            )
+
+            self.test_class_instance.scrubber.register_name_pair(
+                name,
+                self.resource_moniker
             )
         else:
             self.resource = FakeResource(name=name, id=name)
@@ -287,6 +315,8 @@ class ServiceBusNamespaceAuthorizationRulePreparer(_ServiceBusChildResourcePrepa
                                                      client_kwargs=client_kwargs)
         self.parameter_name = parameter_name
         self.access_rights = access_rights
+        if random_name_enabled:
+            self.resource_moniker = self.name_prefix + "sbnameauth"
 
     def create_resource(self, name, **kwargs):
         if self.is_live:
@@ -302,6 +332,11 @@ class ServiceBusNamespaceAuthorizationRulePreparer(_ServiceBusChildResourcePrepa
 
             key = self.client.namespaces.list_keys(group.name, namespace.name, name)
             connection_string = key.primary_connection_string
+
+            self.test_class_instance.scrubber.register_name_pair(
+                name,
+                self.resource_moniker
+            )
         else:
             self.resource = FakeResource(name=name, id=name)
             connection_string = 'https://microsoft.com'
@@ -337,6 +372,8 @@ class ServiceBusQueueAuthorizationRulePreparer(_ServiceBusChildResourcePreparer)
         self.parameter_name = parameter_name
         self.access_rights = access_rights
         self.servicebus_queue_parameter_name = servicebus_queue_parameter_name
+        if random_name_enabled:
+            self.resource_moniker = self.name_prefix + "sbqueueauth"
 
     def create_resource(self, name, **kwargs):
         if self.is_live:
@@ -354,6 +391,11 @@ class ServiceBusQueueAuthorizationRulePreparer(_ServiceBusChildResourcePreparer)
 
             key = self.client.queues.list_keys(group.name, namespace.name, queue.name, name)
             connection_string = key.primary_connection_string
+
+            self.test_class_instance.scrubber.register_name_pair(
+                name,
+                self.resource_moniker
+            )
         else:
             self.resource = FakeResource(name=name, id=name)
             connection_string = 'https://microsoft.com'