+++ sdk/storage/azure-storage-file-share/tests/test_file_client.py
@@ -371,12 +371,7 @@ class StorageFileClientTest(StorageTestCase):
 
         def callback(response):
             self.assertTrue('User-Agent' in response.http_request.headers)
-            self.assertEqual(
-                response.http_request.headers['User-Agent'],
-                "azsdk-python-storage-file-share/{} Python/{} ({})".format(
-                    VERSION,
-                    platform.python_version(),
-                    platform.platform()))
+            assert "azsdk-python-storage-file-share/{}".format(VERSION) in response.http_request.headers['User-Agent']
 
         service.get_service_properties(raw_response_hook=callback)
 
@@ -389,23 +384,19 @@ class StorageFileClientTest(StorageTestCase):
 
         def callback1(response):
             self.assertTrue('User-Agent' in response.http_request.headers)
-            self.assertEqual(
-                response.http_request.headers['User-Agent'],
-                "TestApp/v1.0 azsdk-python-storage-file-share/{} Python/{} ({})".format(
+            assert ("TestApp/v1.0 azsdk-python-storage-file-share/{} Python/{} ({})".format(
                     VERSION,
                     platform.python_version(),
-                    platform.platform()))
+                    platform.platform())) in response.http_request.headers['User-Agent']
 
         service.get_service_properties(raw_response_hook=callback1)
 
         def callback2(response):
             self.assertTrue('User-Agent' in response.http_request.headers)
-            self.assertEqual(
-                response.http_request.headers['User-Agent'],
-                "TestApp/v2.0 azsdk-python-storage-file-share/{} Python/{} ({})".format(
+            assert ("TestApp/v2.0 TestApp/v1.0 azsdk-python-storage-file-share/{} Python/{} ({})".format(
                     VERSION,
                     platform.python_version(),
-                    platform.platform()))
+                    platform.platform())) in response.http_request.headers['User-Agent']
 
         service.get_service_properties(raw_response_hook=callback2, user_agent="TestApp/v2.0")
 
@@ -416,15 +407,12 @@ class StorageFileClientTest(StorageTestCase):
 
         def callback(response):
             self.assertTrue('User-Agent' in response.http_request.headers)
-            self.assertEqual(
-                response.http_request.headers['User-Agent'],
-                "azsdk-python-storage-file-share/{} Python/{} ({}) customer_user_agent".format(
+            assert ("customer_user_agent azsdk-python-storage-file-share/{} Python/{} ({})".format(
                     VERSION,
                     platform.python_version(),
-                    platform.platform()))
+                    platform.platform())) in response.http_request.headers['User-Agent']
 
-        custom_headers = {'User-Agent': 'customer_user_agent'}
-        service.get_service_properties(raw_response_hook=callback, headers=custom_headers)
+        service.get_service_properties(raw_response_hook=callback, user_agent='customer_user_agent')
 
     @GlobalStorageAccountPreparer()
     def test_error_with_malformed_conn_str(self, resource_group, location, storage_account, storage_account_key):