+++ sdk/eventhub/azure-eventhub-checkpointstoreblob/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Release History
 
+## 1.1.1 (Unreleased)
+
+
 ## 1.1.0 (2020-03-09)
 
 **New features**